Fix forward paragraph motion when current line has length 1
Suppose there are three lines in a file. The first line has one
character — a non-whitespace character. The other two lines are blank.
If the user positions the cursor on the first line and presses }, the
cursor will move to the third line. That is wrong. The cursor should
move to the second line.